requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=3.0
psycopg2-binary>=2.9
python-dotenv>=1.0
Flask-Migrate>=4.0
uvicorn>=0.22
uvicorn[standard]>=0.23
